220506_장고_프로필수정/비번변경/회원탈퇴 기능 추가

+++ meokjalal/taste/review/views.py
@@ -12,7 +12,6 @@ from .forms import CommentForm, ReviewForm
 
 
 def review_list(request):
-
     reviews = Review.objects.all().order_by('-pk')
     comment_form = CommentForm()
     comments = Comment.objects.all()
@@ -153,9 +152,18 @@ def review_edit(request, review_pk):
 @login_required
 # @require_POST
 def review_delete(request, review_pk):
-    review = get_object_or_404(Review, pk=review_pk)
-    print(request.user)
-    print(review.author)
-    if review.author == request.user:
-        review.delete()
-    return redirect('review:review_list')
+    # Ver1. 확인 페이지를 거친다면?
+    if request.method == 'POST':
+        review = get_object_or_404(Review, pk=review_pk)
+        print(request.user)
+        print(review.author)
+        if review.author == request.user:
+            review.delete()
+            return redirect('review:review_list')
+    return render(request, 'review/review_delete.html')
+
+    # Ver2. js (confirm?)으로 확인할거라면?
+    # review = get_object_or_404(Review, pk=review_pk)
+    # if review.author == request.user:
+    #     review.delete()
+    #     return redirect('review:review_list')
